+++ tumitfahrer/TimelineViewController.m
@@ -65,7 +65,7 @@
 
 -(void)viewWillAppear:(BOOL)animated {
     [super viewWillAppear:YES];
-    self.screenName = [NSString stringWithFormat:@"Timeline screen: %d", self.index];
+    self.screenName = [NSString stringWithFormat:@"Timeline screen: %d", (int)self.index];
 
     [self.delegate willAppearViewWithIndex:self.index];
     [self.tableView reloadData];
